moved PubData and BrewData to /models/data
