+++ src/main/java/org/launchcode/brewpub/models/data/PubData.java
@@ -1,4 +1,6 @@
-package org.launchcode.brewpub.models;
+package org.launchcode.brewpub.models.data;
+
+import org.launchcode.brewpub.models.Pub;
 
 import java.util.ArrayList;
